Reduce flicker by disabling subsystems
EPS32 core runs RTOS and wants to service it's hardware perhiperals at regular intervals. Unfortunately those intervals conflict with the strict timing requirements for sending data over the neopixel bus. This commit does the following:
* Disable Bluetooth
* Disable Wifi
* Use the delay function with a 1ms delay to ensure that strip.show() happens right after the interupt cycle, giving maximum opportunity to send as many bytes as we can before the next interrupt happens.
* Disabling the touch_pad actually seems to make it worse, so that is commented out.

Already present from an earlier commit, we are also using portDISABLE_INTERRUPTS() and portENABLE_INTERRUPTS() to try and prevent interrupts happening when we send to the neopixels. 

All this helps a lot, but there is more we can do here.